refactor: ai_report 파일 분할
